incorporate memset0 into into apply_learning_with_momentum
apply_learning_with_momentum() was always followed by a memset
resetting the delta array, which arguably might as well be in the
function, simplifying at the calling site.  The clincher though is
that the main apply_learning loop is a tangle of temporal dependency
and the processor really wants to have something independant to do.
Zeroing the deltas is perfect, and the memset becomes something like a
no-op. ~ 4% speed-up on test-backprop.