recur-nn: wrap and optionally replace clbas_saxpy
cblas_saxpy is likely* to respect de-normal numbers, thereby slowing
down by a few orders of magnitude when they occur.  The quick thing to do
is to denormals down to zero, as is enabled by -ffast-math, which recur-nn
shold be compiled with.  There are probably** no algorithmic optimisations
available to BLAS to accelerate scaling over a naive but alignment aware
implementation, so a local function is likely to be faster than BLAS if
there are denormals around, as is likely*** in momentum arrays.

But it hasbn't been tested.

*That is, unless the BLAS has been compiled with -ffast-math system wide.

**Where "probably" means "nothing that I can imagine".

***This "likely" is pure speculation.