Don't zero delta matrices in apply_learning_with_momentum
The argument for zeroing is that the deltas are going to be zeroed soon
after and zeroing saves a separate memset.  But in some cases (e.g.,
synchronic batching) the deltas are not zeroed, and zeroing loses
information for e.g. pgm-dumps.

The memsets are added in a couple of places where they might be needed.