Add lapped validation and restart training at previous spot
If a training session is repeatedly interrupted, it previously
restarted each time from the beginning of the text, overtraining on
those passages. Now by default it starts from the point in the epoch it
stopped at, though that can be overridden.

The lapped vaildation splits the validation set into smaller parts, which
are tested in successive validation runs and averaged. Each result is for
the whole validation set, though it uses different generations of the net
for different parts.  This saves time over validating the whole lot, but
the result is almost as good and quite smooth.