rnnca: add offsets property, remove offsets constants
Now you can choose the offsets by setting the offsets property to a
special string. The following snippet

  offsets=Y00011112C000111

is parsed as ( <luma-or-chroma [YC]> ( <offset 1[0-9]> <offset 2[0-9]> )*)*

where the Y or C indicates whether following digits are for luma or chroma
(Cb, Cr) inputs. Digits occur in pairs, and indicate offsets that will be
rotated and reflected into 1, 4, or 8 offset positions.  For example, "00"
points to the single offset (0, 0). "01" expands into (0, 1), (1, 0), (0, -1),
and (-1, 0). "12" expands into 8 offsets, with all combinations of positive
and negative 1s and 2s in both positions.  Thus "Y000111C12" indicates the
luma (Y) inputs use all pixels a king's move away, as well as the origin;
while the chroma (Cb and Cr) inputs use the 8 squares a knights move away.