gst-classify: experimental error-weights property
The idea is that when you have many training examples of one class, and
few of another, it is more important to get the rare ones right, so the
errors there should be amplified.

For example in a 3 class system where the second class is rare in the
training data, the following should make its error 3 times as weighty:

   ... ! classify error-weight=1:3:1

This might not actually work.