Adaptive min BPTT error, on by default in text_backprop and gstclassify
The BPTT cycle runs until either
 a) it hits maximum depth (typically 30),
 b) the backpropagated error exceeds some value (curbing explosions)
 c) the backpropagated error drops below some value below which it is
    deemed pointless to continue (i.e., vanishing gradient).

In the case of (c), the cutoff value needs to vary according to the task,
architecture, and initial error (and possibly learning rate) but it is
tricky to know what is a good value.

This new method adjusts the threshold to try and maintain the use of
3/4 of the maximum depth.  This allows some headroom for examples with
good error backpropagation to continue further than average, while
getting the most out of average tasks.

There is still an absolute minimum error, in case it starts to drift into
subnormality.