test_backprop: option to predict capitalisation
Capitalisation prediction in parallel with character prediction, using the
same network.  There is one extra input, flagging a capital, and an extra
softmax pair of outputs, predicting a capital.

Here's an example to show what I mean by working in parallel. Supposing
the text comes to a point that seems to want a person's name, and the
most likely names, given recent history, are "Alice" and "Bob". The
pronouns "he" or "she" could work here.  Accordingly the character
predictor gives high probabilities to "a", "b", "s", and "h", and the
capitalisation predictor gives a relatively high probability, say 50%,
to a capital. It can't say "uppercase for A and B, lowercase for s and h".
Thus a confabulation based on sampling is likely to end up trying to start
with a capitalised pronoun, e.g. "She". In the next generation, reflecting
on the "S", it might try to recover by aiming for, say, "Sam", but that is
another matter.

The reported entropy is for the character predictions only. Having the
capitalisation information seems to result in slightly better character
predictions (cross entropy of ~1.984 vs 1.992, after 5M cycles, 200
hidden nodes).

This was somewhat invasive and involved a bit of tidying, but it doesn't
seem to have broken anything.