recur-nn: default to naive matrix ops in opinion, quicker than cblas
The naive version can skip a row when an input is zero.  In a newly
initialised net, this means roughly half the hidden nodes, and in
trained nets, seemingly more.

For one-hot representations, most of the input nodes are zero, but
other feature vectors are rarely zero.

The testing that found the naive operations to be quicker was carried
out on the one-hot test_backprop task, so the benefits will be smaller
on MFCC tasks.