aligned alloc routines use malloc/calloc if they are aligned enough
This is a bit of a hack. On 32 bit machines, glibc malloc allicates on
64 bit boundaries, while on 64 bit machines it uses 128 bit
boundaries. 128 bit boundaries are what we want.

This doesn't seem to be any quicker (which is allegedly possible), but
it potentially uses less memory in the calloc case.  If the memory is
zeroed in userspace (e.g. by memset) all the pages are dirtied and the
memory is really allocated even if it isn't eventually used.  Calloc
can promise zeroed pages which don't actually have to be zeroed until
they are used.  This should save a lot in cases where, for example, a
BPTT struct with momentum and delta arrays is allocated but these
aren't actually used.