gstclassify: shift some temporary numbers into pending GValues
These properties have to be set before the net is created, and can't be
changed afterwards. Because there is a little gap of a few milliseconds,
they have to be stored somewhere, and have until now been cluttering up
the GstClassify struct. Now they are saved in the pending_properties list.

It saves a little bit of space.