trying out a different procedure for summing delta
The idea is that the delta arrays for each training net are transient,
so they might as well be dealt with and summed immediately, then their
memory can leave cache. (in fact, there only needs to be a single
delta array shared by all instances, but this commit doesn't go that far).

The result appears to be equivalent, but is arrived at slightly faster
according to very simple testing.