recur-nn: rearrange Nesterov momentum to avoid preparatory call
The formula for Nesterov momentum appears to ask for the momentum to be
added to the weights before the momentum update, but you regard the first
update to be zero (as it is), and ignore the last one, then doing the
adding afterwards amounts to the same thing.  This is the same kind of
logic that led to simplified Nesterov momentum, but the result is slightly
different, I think.

This could all be wrong: I haven't analysed it deeply. Try and see.