Makefile: remove broken dependencies on path.h
For some reason, the elf files (i.e. linking) were depending on path.h,
which was only ever needed to create the object files (compiling).
This caused the inclusion of "precompiled headers" in the ELF, and made
test_backprop refuse to compile while an existing version of it was
running.

Unfortunately, the files that do need path.h can't get it via the magic
of -MMD. See http://gcc.gnu.org/bugzilla/show_bug.cgi?id=57951.