clasify: add learn-rate-scale properties, and revamp property functions
top-learn-rate-scale changes the learnning rate for the top layer
relative to the default rate.

bottom-learn-rate-scale does the same for the bottom layer learn rate, if
there is a bottom layer.

This prompted a partial rewrite of the communication between gstreamer
properties and net attributes.  Now more attributes are writable at
run-time, and more properties are readable.