Shift bottom layer flag back to original value
When the accumulator flag went I shifted the bottom layer flag down
to the accumulator flags place.  That made the old nets that had
accumulators wrongly think they had a bottom layer when reloaded.

Eventually the accumulator flag can be replaced with something new, but
not just yet.

Also, try to cope with it in recur-nn-io.