classify: adjust to multi-group classification via heinous rewrite
Besides changing the syntax for communication between Python and
gstreamer, the use of multiple groups contravenes assumptions made on both
sides.  This has been an opportunity for various tidy-ups, for example
removing the broken microphone input option, and removing the concept of
untimed files.

It is not clear whether this actually works, in part because logging seems
to be broken.