classify: properly manage circular audio buffer; improve terminology
Valgrind pointed out that the audio buffer was sometimes being used
uninitialised, which turned out to be due to the buffer being read ahead
of where it should have been.  The read pointer needs to lag behind
the write pointer sufficiently for the a complete chunk (a half window for
each channel) to be read. Instead it was getting to overlap or equal the
write pointer.  This should have lead to dreadful performance.

Part of the problem, I think, is that it was not very clear from the
names how it should have worked.