classify: allow audio queue to be misaligned with chunk size
The audio queue size has always been an exact multiple of the chunk
size, so up until now there was no chance of the reader overrunning
the queue. Offset randomisation means the reader will overrun, and the
simplest thing to do seems to be to extend the memory allocation a bit,
and copy the beginning of the queue in when necessary.