Add metadata attribute to nets, as yet unused and untested
The classify plugin uses a wide variety of settings outside of the net
itself.  For example, the audio frequency limits and knee are set
differently for different applications.  It is easy to mistakenly load
a net trained on a range of say 100 to 3900 Hz into a classifier using
a range of 300 to 3900 Hz, and to wonder why it doesn't work.

If the classifier compiles metadata about its settings and saves it
with the net, it can compare its current metadata with that saved and
warn or abort if it doesn't match. Eventually it should be possible to
load mtadata directly from the net and not have to guess a correct
commandline.