recur-nn-io: Don't save/restore bptt arrays and other temporary state
The original conceit was that you could save and restart a net, and it
would continue *exactly* as if you hadn't stopped it. While that was
neat, it has never been used, and the promise has gradually been
eroded, for instance by the bottom layer momentums not being saved.

So rather than half-heartedly not quite storing full state, it seems
clearer to save the learnt weights and the dimensions and
metaparameters of the temporary arrays. This gives just as good
performance in practice, and reduces the size of the saved nets by
about a third (while presumably speeding up IO by a related factor).

What is lost is the theoretical possibility of a range of precise
tests. If those tests ever seem like a reality, saving and loading
temporary state can easily be restored.