Classify: add the option to use derivative features
Many speech recognition papers use FFT bins and the first and second
derivatives.  This will do that if you set the 'delta-features'
property to 2.  If you set it to 1 or 3 it uses just the first, or the
first three derivatives, respectively.

Unfortunatey, it breaks the loading of nets, because the metadata
won't match.

Untested.