classify: kiwi-* load classes from file names, and otherwise fit TMP task
The classes in the TMP training data are encoded in the filenames (first
letter stands for the class).  The --classes-from-file-names option does
that (instead of -t/--timings).

quite a bit of common code has been moved from kiwi-train and kiwi-test
into classify.py, and obsolete constants have been removed from
classify_kiwi.py (which is doomed to disappear).