classify: optionally load call timing from json with thresholds
There is a theory that some of the morepork calls are impracticably
faint, so that training on them is counterproductive. We can use an
estimate of call intensity (calculated elsewhere) to ignore some of
the faint calls.

The ignored calls are not non-moreporks though, so we need to use the
non-training state.