classify: schedule learn-rate using exponential decay based on generation
Before the learn-rate descended a series of explicit steps from the time
kiwi-train was started.  This was OK until a training run was stopped and
restarted, at which point it would begin again at the highest learn-rate.

Now the learn rate depends on the generation, and will gradually decay
until it hits the minimum value.

The same system is used for dropout, but there the minimum value is
followed by zero dropout.

Lightly tested.