classify: optionally load net earlier, allowing generation to be read
Also, use the start() hook to create the net, instead of using
setup() with a guard.

The exponential learn-rate mechanism sets the learn-rate based on the
current generation, but for the first cycle the net is not loaded before
the learn-rate is set, leading to a false generation reading of zero,
thence a bad learn-rate.  This adds a magic property that triggers the
loading of the net, so the trainer can make sure it happens in time.