classify: kiwi-test shows presence/absence ROC curves
Thes work by sorting the scores for the minute (skipping the first 10
samples) and taking the n-th highest, where n is somewhere between 0
and 19. The theory is the highest number might be a flukish transient
peak, while the n-th highest is either a genuine score or n transient
peaks.  Somewhere between the 5th and 9th highest value looks good.