recur-nn: use equality in min_error test to catch zero threshold
When the incoming error sum is zero (i.e, when there is no training
information, or the response was perfect), min_error_sum is zero, and
error_sum is never less that it. On the other hand, it is never greater
than it, and completing the BPTT cycle is obviously pointless. With the
<= test, the short-circuit break works.