recur-nn-init: simple loop-oriented initialisation
This tries to expicitly establish long term dependencies by forming
loops of a particular size that maintain a particular gain.

In the test_backprop case, it makes 108 loops of length 20, each with
a net gain of 0.2.  That is, all things being equal, input into a node
on the loop will arrive back there 20 generations later at 1/5th the
power.  The feedback might be negative or positive.

The effect is not great for test_backprop, where (presumably) loops of
20 are too long to be useful and the balance is all mucked up. But
this is just the beginning.