init: add a maximally disconnected initialisation mode
The loops and runs initialisation modes arrange things so that each
weighted edge leads to another weighted edge.  The disconnected mode
tries the opposite: to the extent practical*, it ensures each weighted
edge leads to no weighted edges, and that each input weight points to
an unused node.

* "to the extent practical" means "not very much at all", if the
number of weighted edges is greater than half the number of hidden
nodes.  In fact what happens in the case where the weighted edges
significantly exceed the number of nodes is the arrangement of paths
becomes more even than would be the case with random connections, with
each node having roughly the same number of incoming and outgoing
connections.