Attempt clockwork-RNN, as per Joutnik et al -- possibly broken
That's arXiv:1402.3511v1 [cs.NE]

Not all the weights are used in each round, which means not all the error
is back-propagated each bptt cycle. As in that paper, this uses powers
of two as the clock cycles.

There is a new clock counter for the opinion, because net->generation
is incremented by training only.

There are likely to be mistakes and misunderstandings.

It seems that clockwork is not particularly useful for test_backprop,
where the immediate history is very important. It seems to be at least
stable for the morepork task.