rnn core: refurbish rnn_opinion, reduce unnecessary operations
There was a small loop to zero the tail of the hidden_layer in case it
contained left over remnants of an input layer -- in the case where the
input and hidden layers were being used alternately to save copying. In
fact this alternation is never used, so the zeroing was pointless.

Also, defererence net->hidden_layer just once, in case it helps.