rnn, test_backprop: add periodic gaussian noise
This form of regularisation that works well in RNNs (sorry, can't remember
references), because it doesn't draw the weights towards zero (or any
other specific value) so the recurrent dynamics are preserved. On the other
hand it is very slow.

A test_backprop with 100 neurons, 1M generation, 4k intervals,
learn-rate 1e-3, arrives at almost equivalent entropy with and without
weight noise of 1e-4.  Without weight noise it takes ~50 seconds; with it,
~400 seconds.