test_backprop: gaussian noise was happening every iteration
Now the previous slowness makes sense.

Adding noise every 4k to test_backprop with 100 neurons increases the time
taken by roughly half a percent.

With a learn-rate of 1e-3, noise rates of 3e-3 or less give results no
worse than no noise. No better, either, but that is to be expected with
short tests, because overfitting is not an issue.