rnn: separate bptt loop break threshold from scaling threshold
This gives the bptt loop a chance to sort itself out, while allowing the
tighter constraints on out of control learning.

Actually, if the bptt gain goes above and then below the threshold, the
net gain will be greater than the threshold without the scaler noticing,
which may be bad.