classify: cleverer presence detection in kiwi-test
The old method (which is still available) chose the kth highest score of
each file, where k is selected empirically/magically according to the
species. The idea was that the highest scores might be transient, but
actual birds would raise the score for a number of windows. It worked ok.
With kiwis, because their calls last 20 seconds, it was reasonably to look
at, say, the 80th window at 15.6 windows per second -- a real call would
affect more than 80 windows but a glitch wouldn't.

The new way looks at a rolling average of the instantaneous scores, and
returns the highest (by default). This eradicates short glitches and
prevents a whole lot of brief errors adding up to a bad result. Limited
testing with kiwis suggests a noticable increase in presence AUC.

The instantaneous scores can be downsampled (box filter) to save a little
bit of time with hopefully little ill effect. For example, this:

  --presence-subsample=1 --presence-run-length=15 \
	--presence-ignore-start=1

collects the results into roughly 1 second blocks, then calculates a
running average of 15 of them. Without `--presence-subsample`, the running
average would be of 15 * 15.6 ~ 230 scores.