rnn: maybe_scale_hiddens seems unnecesssary, adjust input scaling
Scaling the hiddens is unnecessary for the ReLU top layer, and is
redundant for the recurrent layer, given that the inputs get scaled
in the next round.

This enthusiastic clamping comes from when this was an always learning
artwork -- for most application a haywire net can just be dumped or
rewound and this somewhat damaging care is unnecessary.