test_backprop: remove Gutenberg text from Erewhon test file
test_backprop uses the last little bit of the file as validation text.
The last little bit of Erewhon was a Project Gutenburg ramble, which
was not very representative of the text as a whole, and thus not all
that good for validation.

The previous default text was shuffled, which made the validation text a
reasonable sample. Shuffling Erewhon would also be possible, but then
you lose an interesting book from your repository.

The Project Gutenberg text contained license terms, and it would normally
be wicked to delete that, but Erewhon is out of copyright worldwide, so
the license is inapplicable, and the Gutenberg text pretty well said as
much.

* also adjust the default character set accordingly.