test_backprop: rework metadata to use simple separators
sscanf() doesn't work for test_backprop metadata because the alphabet
and collapse_chars fields can (and usually will) contain spaces. I looked
at using json, xml, or ini formats, but they all seemed like a lot of work
for little immediate gain.

In the end I just used the ASCII unit separator (\x1F), which is unlikely
to turn up in any alphabet, and was once upon a time meant for this
purpose.  There could of course be trouble with this later.