text-classify: try to cope with TEI <choice> elements
In the TEI xml format, <choice> elements contain two or more versions of
the same thing. In the NZETC texts, they seem to usually contain a
<orig/><reg/> pair, where <orig> holds a name as written, and <reg> holds
a regularised version. Here is an example (slightly reformatted):

<name type="person"><choice>
 <orig><foreign xml:lang="en">Rev. M</foreign>. Pahewa</orig>
 <reg><foreign xml:lang="en">M</foreign>. Pahewa</reg>
</choice></name>

A naive reading doubles the name in the text, so with this patch we drop
all children except <orig>.

This is not really correct, as <choice> can contain any number of elements
from this set:

   abbr choice corr expan orig reg sic unclear

See http://www.tei-c.org/release/doc/tei-p5-doc/en/html/ref-choice.html