Add optional pre-synaptic noise
This adds gaussian noise after the matrix calculation, and before the
non-linearity (i.e. the rectifier). On test_backprop, the effect seems to
be to speed up each generation, but to converge at a higher entropy than
would be obtained without noise.

Not very well tested outside of test_backprop.