Move Gstreamer plugins to their own directory
Build `libgst*.so` in `./plugins`, not `.`. The practical advantage of
this is saving hundreds of milliseconds at the startup of each
gstreamer-using process. At start up, gstreamer searches all the
directories in its plugin path for plugins. When `.` is in the path,
gstreamer looks through all the subdirectories including, e.g.
net/autosave which can end up with thousands of files.

If there are cases where the plugin directory is still set to `.` in a
player, the plugins will still be found due to the recursive search.