test_backprop: add ADAGRAD options
The momentum arrays is used to accumulate the gradient squared magnitude,
and as this is used in division it has to start at a non-zero value. A
low initial value causes the effective learning rate to initially drop
quickly and possibly wildly, while a higher one keeps it steady but less
responsive.  A similar parameter might be used in other adaptive learning
mechanisms. Hence --ada-ballast, describing the inertia effect while not
tying it to adagrad specifically.