charmodel: use a struct for alphabet; tug at test_backprop knots
In several places, the alphabet and collapse_chars arrays were being
passed around with their lengths, and the functions signatures all
contained this:

    int *alphabet, int a_len, int *collapse_chars, int c_len,

Now there is a struct that contains all that, simplifying stuff here and
there.

This began as an attempt to further untangle test_backprop, and at least
the rnn_char_check_metadata() function has come out of load_or_create_net()
into the charmodel library.

text-classify has not been keep up to date, since it doesn't work
properly anyway. This is symbolic punishment for it.