New program that prints cross entropy of files vs model
text-cross-entropy predicts the contents files named on the command line
and prints the corresponding cross entropy. Typical usage:

    find [...] | xargs ./text-cross-entropy -f some.net -m 200 -i 10

which will look at files longer than 200 bytes, ignoring the entropy
of the first 10 letters, and print something like this:

    a.txt 2.727
    b.txt 1.8914
    [...]

The `-p` option pretends that each file begins with the given string.
Combined with `-i` which ignores the start, this allows the net to smooth
over the sudden shock of starting from an empty state.