py-charmodel: helper functions for getting/setting net/bptt parameters
This is a little clumsy, but seems to be better than the alternative
of having a dedicated function for each one.

It would probably be possible to do this without the dereference of
(void *)closure into an (int *) offset by instead storing directly the
(int)offset in (void*)closure. The closure would be a zero-page pointer,
which would be entirely fine if python never used it as such. But who
knows with python.