py-charmodel: add classmethod to load a net
This involves quite a lot of rearrangement -- taking common code out of
Net_init so the new load system can use it too. Alphabet is also changed
to work with simple alphabet and collapse_char strings rather than
recalculating it from a text. That is so the alphabet can be reconstructed
from the net metadata. Python needs to pass in a metadata decoder (as
it has to encode it in the first place). A simple example is given in
test/charmodel.py (along with a test net), where the metadata is stored
as json.

Why can't Net_init do it all, loading a net if, say, no alphabet is given
but a filename and metadata decoder is? Well, it could (and one day might)
but it would be complicated.