py-charmodel: propagate metadata parsing exceptions
The metadata is parsed by a passed-in Python function. If an error
occurs in the Python, the result will be NULL and an exception will be
set up. Before we ignored that and tried to use the NULL; now we let
the exception bubble up to whoever wants to catch it.