Add text-predict --diagonal-only-section and related options
This restricts a portion of the recurrent neurons to only listen to
themselves and the inputs, somewhat similar to the structurally
constrained recurrent network (SCRN) proposed by Mikolov et. al in
arXiv: 1412.7753. A difference is that they use linear activations on
the diagonal units.

--diagonal-only-friends leaks one hidden node into each "diagonal-only"
node.

--diagonal-only-boost alters the initial weights by a constant factor.
The idea is that you probably want them to be quite high and all or
mostly positive.

This is a slightly hackish proof of concept.