classify: add an option to write raw features to a file
It is said that training works better if the input features are well
normalised, which we haven't done much so far. In order to normalise
the inputs we need to know what normal values look like.

For the bird song task the volume of calls and background noise is
quite unpredictable (birds move, wind blows, etc), so good
normalisation seems difficult. That's why I ignored it. But for the
radio language recognition task the audio falls within a narrower
range.