classify: add features-offset and features-scale parameters
With `classify-train --features-offset` you can shift the feature values,
to (for example) zero their means. Similarly, you can multiply the
weights by constants with `classify-train --features-scale`. The offset
is applied first.

These take colon separated float values, applied to as many features as
there are numbers, starting from the first. For example, the following
would adjust the scale of the first 5 features:

   classify-train --features-scale 1.234:-2.2e5:24:1E-12:0.00021

Not well tested, but it seems to work.