classify: add --balanced-training option
With --balanced-training=1, the chance of an example being ignored is
proportional to the number of examples of that class seen so far.

With a value between 0 and 1, common classes are favoured somewhat, so
that the training occurs more often on the common class, but less
often than would ordinarily be the case.

With a value greater than 1, rare classes are overly favoured, so that
it looks to the net as if they are the most common.

With --balanced-training=0, nothing is done, as if you didn't use
--balanced-training at all.

It seems that a value slightly below 1 might be best for bird call
detection where one class is quite rare.