classify: fully adaptive auto-save, using decaying records
Previously there were some very ad-hoc estimates of what constituted a
good result; some tasks would save always while others would very
rarely. This makes a save occur when a net is as good as or better
than recent ones according to various metrics. It causes a lot of
saves early on (which the old method didn't), and fewer but regular
ones later.