classify_stats: calculate accuracies as well as AUC
Two definitions of accuracy are used. "DFD" is the maximum distance from
the y=x diagonal on the ROC curve. I think it roughly corresponds to the
macro-average of accuracies for the various classes. On the other hand,
"accuracy" refers to the maximum micro-average.  In other words, "DFD" is
the accuracy you would get if all the classes occurred in equal numbers,
while "accuracy" is what you get from the actually occurring proportions in
the test set.

Also, because "DFD" is thinking in terms of the ROC curve, the maximum is
0.707, while "accuracy" can go up to 1. That might change.